+++ lib/screens/homePage.dart
@@ -191,7 +191,7 @@ class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      automaticallyImplyLeading: false, // Oculta la flecha de retroceso
+      //automaticallyImplyLeading: true, //RLP Oculta la flecha de retroceso
       elevation: 0,
       backgroundColor: const Color.fromARGB(255, 231, 232, 234),
       title: Row(
